# backend/openapi.yaml
openapi: 3.1.0
info:
  title: VINK Group API
  version: "1.0.0"
  description: |
    Backend API for the VINK Group corporate site. Two categories of route:
    public content reads + public form submissions (rate-limited), and three
    admin-only reads (bearer-token auth, also rate-limited).
servers:
  - url: https://vink-holding-backend-production.up.railway.app/api
    description: Production
  - url: http://localhost:4000/api
    description: Local dev

components:
  securitySchemes:
    adminSession:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        A session token issued by POST /admin/login (8h expiry by default). Requires
        ADMIN_PASSWORD_HASH and JWT_SECRET to be configured on the server; unset on
        the server -> 503, not open access.
  schemas:
    Error:
      type: object
      properties:
        error: { type: string }
      required: [error]
    Sector:
      type: object
      properties:
        id: { type: string }
        name: { type: string }
        shortDescription: { type: string }
        statLedSubhead: { type: string }
        description: { type: string }
        ctaLabel: { type: string }
        icon: { type: string }
        markets: { type: array, items: { type: string } }
        subsidiaries: { type: array, items: { type: string } }
        stats:
          type: array
          items:
            type: object
            properties:
              label: { type: string }
              value: { type: string }
    NewsArticle:
      type: object
      properties:
        id: { type: integer }
        category: { type: string }
        title: { type: string }
        date: { type: string, format: date }
        excerpt: { type: string }
        body: { type: string }
    Job:
      type: object
      properties:
        id: { type: integer }
        title: { type: string }
        department: { type: string }
        location: { type: string }
        type: { type: string }
        description: { type: string }
    ContactSubmission:
      type: object
      required: [name, email, message]
      properties:
        name: { type: string, maxLength: 200 }
        email: { type: string, format: email, maxLength: 254 }
        phone: { type: string, maxLength: 40 }
        subject: { type: string, maxLength: 200 }
        message: { type: string, maxLength: 5000 }
    ApplicationSubmission:
      type: object
      required: [name, email, jobTitle]
      properties:
        jobId: { type: [string, "null"] }
        jobTitle: { type: string, maxLength: 200 }
        name: { type: string, maxLength: 200 }
        email: { type: string, format: email, maxLength: 254 }
        phone: { type: string, maxLength: 40 }
        coverNote: { type: string, maxLength: 5000 }

paths:
  /health:
    get:
      summary: Health check
      operationId: getHealth
      security: []
      responses:
        "200": { description: OK }

  /company:
    get:
      summary: Company profile
      operationId: getCompany
      security: []
      responses:
        "200": { description: OK }

  /sectors:
    get:
      summary: All six business sectors
      operationId: listSectors
      security: []
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { type: array, items: { $ref: "#/components/schemas/Sector" } }

  /sectors/{id}:
    get:
      summary: Single sector by slug
      operationId: getSector
      security: []
      parameters:
        - { name: id, in: path, required: true, schema: { type: string } }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Sector" }
        "404": { description: Not found, content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } } }

  /markets:
    get:
      summary: Operating markets (7 countries)
      operationId: listMarkets
      security: []
      responses:
        "200": { description: OK }

  /ads:
    get:
      summary: Sliding-adverts promo data
      operationId: listAds
      security: []
      responses:
        "200": { description: OK }

  /leadership:
    get:
      summary: Group leadership + governance
      operationId: getLeadership
      security: []
      responses:
        "200": { description: OK }

  /sustainability:
    get:
      summary: Sustainability pillars
      operationId: getSustainability
      security: []
      responses:
        "200": { description: OK }

  /jobs:
    get:
      summary: All open roles
      operationId: listJobs
      security: []
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { type: array, items: { $ref: "#/components/schemas/Job" } }

  /jobs/{id}:
    get:
      summary: Single role by id
      operationId: getJob
      security: []
      parameters:
        - { name: id, in: path, required: true, schema: { type: string } }
      responses:
        "200": { description: OK }
        "404": { description: Not found }

  /investor-reports:
    get:
      summary: Investor reports list, newest first
      operationId: listInvestorReports
      security: []
      responses:
        "200": { description: OK }

  /investor-reports/{id}/download:
    get:
      summary: Server-generated PDF for a report
      operationId: downloadInvestorReport
      security: []
      parameters:
        - { name: id, in: path, required: true, schema: { type: string } }
      responses:
        "200":
          description: PDF file
          content:
            application/pdf:
              schema: { type: string, format: binary }
        "404": { description: Not found }

  /stats:
    get:
      summary: Headline group stats (employees, markets, etc.)
      operationId: getStats
      security: []
      responses:
        "200": { description: OK }

  /news:
    get:
      summary: News articles, newest first
      operationId: listNews
      security: []
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { type: array, items: { $ref: "#/components/schemas/NewsArticle" } }

  /news/{id}:
    get:
      summary: Single article by id
      operationId: getNewsArticle
      security: []
      parameters:
        - { name: id, in: path, required: true, schema: { type: string } }
      responses:
        "200": { description: OK }
        "404": { description: Not found }

  /newsletter:
    post:
      summary: Subscribe to the newsletter
      operationId: subscribeNewsletter
      security: []
      description: Rate-limited (8/15min per IP). Deduplicates by normalized email.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email]
              properties:
                email: { type: string, format: email, maxLength: 254 }
      responses:
        "201": { description: Subscribed }
        "200": { description: Already subscribed }
        "400": { description: Invalid email, content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } } }
        "429": { description: Rate limited }

  /contact:
    post:
      summary: Submit the contact form
      operationId: submitContact
      security: []
      description: Rate-limited (8/15min per IP).
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/ContactSubmission" }
      responses:
        "201":
          description: Received
          content:
            application/json:
              schema:
                type: object
                properties:
                  message: { type: string }
                  reference: { type: string, pattern: "^VH-\\d{5}$" }
        "400": { description: Validation error, content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } } }
        "429": { description: Rate limited }

  /careers/apply:
    post:
      summary: Submit a job application
      operationId: applyToJob
      security: []
      description: Rate-limited (8/15min per IP).
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/ApplicationSubmission" }
      responses:
        "201":
          description: Received
          content:
            application/json:
              schema:
                type: object
                properties:
                  message: { type: string }
                  reference: { type: string, pattern: "^VH-APP-\\d{5}$" }
        "400": { description: Validation error }
        "429": { description: Rate limited }

  /admin/login:
    post:
      summary: Admin login — issues a session token
      operationId: adminLogin
      security: []
      description: Rate-limited (5/15min per IP).
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [username, password]
              properties:
                username: { type: string }
                password: { type: string, format: password }
      responses:
        "200":
          description: Session issued
          content:
            application/json:
              schema:
                type: object
                properties:
                  token: { type: string, description: "JWT — send as 'Authorization: Bearer <token>' on admin endpoints." }
                  expiresAt: { type: string, format: date-time }
        "400": { description: Missing username/password }
        "401": { description: Invalid credentials }
        "429": { description: Rate limited }
        "503": { description: Admin login not configured on this environment }

  /careers/applications:
    get:
      summary: "[Admin] All job applications"
      operationId: listApplications
      security: [{ adminSession: [] }]
      responses:
        "200": { description: OK }
        "401": { description: Missing/wrong token }
        "503": { description: Admin login not configured on this environment }

  /contact/messages:
    get:
      summary: "[Admin] All contact submissions"
      operationId: listMessages
      security: [{ adminSession: [] }]
      responses:
        "200": { description: OK }
        "401": { description: Missing/wrong token }
        "503": { description: Admin login not configured on this environment }

  /newsletter/subscribers:
    get:
      summary: "[Admin] All newsletter subscribers"
      operationId: listSubscribers
      security: [{ adminSession: [] }]
      responses:
        "200": { description: OK }
        "401": { description: Missing/wrong token }
        "503": { description: Admin login not configured on this environment }